- Add integration tests - Update manual tests
