+++ pkg/ibm_plugin/utils.go
@@ -22,6 +22,7 @@ import (
 	"hash/fnv"
 	"io"
 	"net/http"
+	"os"
 	"reflect"
 	"strings"
 )
@@ -166,4 +167,14 @@ func GetStructHash(s interface{}, fieldsToExclude []string) (uint64, error) {
 		}
 	}
 	return h.Sum64(), nil
+}
+
+
+// Gets subscription ID defined in environment variable
+func GetIBMResourceGroupID() string {
+	resourceGroupID := os.Getenv("PARAGLIDER_IBM_RESOURCE_GROUP_ID")
+	if resourceGroupID == "" {
+		panic("Environment variable 'PARAGLIDER_IBM_RESOURCE_GROUP_ID' is required for testing")
+	}
+	return resourceGroupID
 }
